broke out massive email function from checkout component
